django/educa/requirements.txt
0==0.0.0
arabic-reshaper==2.1.1
asgiref==3.2.10
asyncio==3.4.3
atomicwrites==1.4.0
attrs==21.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorators==2.0.7
dj-database-url==0.5.0
Django==3.1.1
django-braces==1.14.0
django-ckeditor==6.0.0
django-embed-video==1.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-memcache-status==2.2
django-redis==4.12.1
django-taggit==1.4.0
djangorestframework==3.12.2
docutils==0.16
flake8==3.9.2
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
importlib-metadata==4.0.1
iniconfig==1.1.1
Kivy==1.11.1
kivy-deps.angle==0.1.10
kivy-deps.glew==0.1.12
kivy-deps.sdl2==0.1.23
Kivy-Garden==0.1.4
kivy-garden.contextmenu==0.1.0.dev1
kivy-garden.drag-n-drop==0.1.0
kivy-garden.xcamera==2020.613
kivy-garden.zbarcam==2020.524
kiwisolver==1.2.0
Markdown==3.3.4
matplotlib==3.3.2
mccabe==0.6.1
mpmath==1.2.1
numpy==1.19.2
opencv-python==4.4.0.42
packaging==20.9
Pillow==7.2.0
pluggy==0.13.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
PyQt5==5.15.1
PyQt5-sip==12.8.1
pytest==6.2.4
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.3
python-memcached==1.59
pytz==2020.1
pywin32==228
pyzbar==0.1.8
qrcode==6.1
qrtools==0.0.2
redis==3.5.3
reportlab==3.5.56
requests==2.24.0
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
sympy==1.8
tasks @ file:///C:/Users/PC/Desktop/test/tasks_proj
tinydb==3.15.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.25.10
webencodings==0.5.1
whitenoise==5.2.0
xcamera==2020.613
xhtml2pdf==0.2.5
zbarcam==2020.524
zipp==3.4.1
